fix(cli-sub-agent): preserve ACP startup stderr
Drain the ACP stderr reader before reporting an already-exited child so Gemini pre-handshake failures keep their diagnostic text and classify correctly.

Split ACP connection status and event streaming helpers into dedicated modules so connection.rs remains under the pre-commit file budget.

Verified with CARGO_BUILD_JOBS=2 just pre-commit.
